# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notes/user/{userDid}": {
      "get": {
        "operationId": "NotesController_getNote",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaType",
            "required": true,
            "in": "query",
            "description": "Media type (movie, show, season, episode)",
            "schema": {
              "type": "string",
              "enum": [
                "movie",
                "show",
                "season",
                "episode"
              ]
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "query",
            "description": "TMDB movie ID or show ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seasonNumber",
            "required": false,
            "in": "query",
            "description": "Season number for season/episode items",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "episodeNumber",
            "required": false,
            "in": "query",
            "description": "Episode number for episode items",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a note for a user and media item",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes/user/{userDid}/notes": {
      "get": {
        "operationId": "NotesController_getUserNotes",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to return (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return per page",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notes retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotesResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get paginated notes for a user",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes": {
      "post": {
        "operationId": "NotesController_upsertNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note upserted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or update a note",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes/{noteId}": {
      "delete": {
        "operationId": "NotesController_deleteNote",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a note",
        "tags": [
          "notes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Opnshelf API",
    "description": "Personal media tracker powered by AT Protocol",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "NoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rkey": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "number"
          },
          "episodeNumber": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "rkey",
          "content",
          "mediaType",
          "mediaId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserNoteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "number"
          },
          "episodeNumber": {
            "type": "number"
          },
          "mediaLabel": {
            "type": "string",
            "description": "Human-readable label for the media. Display only — never build a URL from it."
          },
          "mediaTitle": {
            "type": "string",
            "description": "Title of the movie or show that mediaId identifies. Never composite — this is what URL slugs are built from (ADR 0023)."
          },
          "posterPath": {
            "type": "string",
            "description": "Poster path for the movie or show"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "mediaType",
          "mediaId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedNotesResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total count of items across all pages"
          },
          "page": {
            "type": "number",
            "description": "Current page number after server-side clamping"
          },
          "pageSize": {
            "type": "number",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of available pages"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether a next page exists"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether a previous page exists"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserNoteDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "items"
        ]
      },
      "UpsertNoteDto": {
        "type": "object",
        "properties": {
          "mediaType": {
            "type": "string",
            "description": "Media type",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string",
            "description": "TMDB movie ID or show ID"
          },
          "seasonNumber": {
            "type": "number",
            "description": "Season number for season/episode items"
          },
          "episodeNumber": {
            "type": "number",
            "description": "Episode number for episode items"
          },
          "content": {
            "type": "string",
            "description": "Note content",
            "maxLength": 5000
          }
        },
        "required": [
          "mediaType",
          "mediaId",
          "content"
        ]
      },
...
    }
  }
}